+++ CTSTools.WEB/App/Features/Engineering/ComponentID/AttributeManagement/SubClass/SubClass_Controller.cs
@@ -1,4 +1,5 @@
 ﻿using CTSTools.BLL.Common;
+using CTSTools.BLL.Features.AdvancedSettings.SecurityManagement.Action;
 using CTSTools.BLL.Features.Engineering.ComponentID.AttributeManagement.SubClass;
 using CTSTools.WEB.App_Start;
 using DevExtreme.AspNet.Data;
@@ -32,9 +33,12 @@ public class SubClass_Controller : ApiController
     [Route("api/SubClass/Create")]
     public IHttpActionResult CreateSubClass([FromBody] SubClassDTO SubClassDTO)
     {
-
-        SubClassDTO.AddedByID = Auth_Helper.GetLoggedUserOid();
-        var _validationResultDTO = SubClass_Service.CreateSubClass_Global(SubClassDTO);
+        var _validationResultDTO = Auth_Helper.ValidatePermission_Global(nameof(SubClass), (int)Action_Enum.Create);
+        if (_validationResultDTO.Result)
+        {
+            SubClassDTO.AddedByID = Auth_Helper.GetLoggedUserOid();
+            _validationResultDTO = SubClass_Service.CreateSubClass_Global(SubClassDTO);
+        }
         return Json(_validationResultDTO);
     }
 
@@ -42,15 +46,23 @@ public class SubClass_Controller : ApiController
     [Route("api/SubClass/Update")]
     public IHttpActionResult UpdateSubClass([FromBody] SubClassDTO SubClassDTO)
     {
-        SubClassDTO.LastUpdateByID = Auth_Helper.GetLoggedUserOid();
-        var _validationResultDTO = SubClass_Service.UpdateSubClass_Global(SubClassDTO);
+        var _validationResultDTO = Auth_Helper.ValidatePermission_Global(nameof(SubClass), (int)Action_Enum.Update);
+        if (_validationResultDTO.Result)
+        {
+            SubClassDTO.LastUpdateByID = Auth_Helper.GetLoggedUserOid();
+            _validationResultDTO = SubClass_Service.UpdateSubClass_Global(SubClassDTO);
+        }
         return Json(_validationResultDTO);
     }
     [HttpPost]
     [Route("api/SubClass/Delete")]
     public IHttpActionResult DeleteSubClass([FromBody] SubClassDTO SubClassDTO)
     {
-        var _validationResultDTO = SubClass_Service.DeleteSubClass_Global(SubClassDTO);
+        var _validationResultDTO = Auth_Helper.ValidatePermission_Global(nameof(SubClass), (int)Action_Enum.Delete);
+        if (_validationResultDTO.Result)
+        {
+            _validationResultDTO = SubClass_Service.DeleteSubClass_Global(SubClassDTO);
+        }
         return Json(_validationResultDTO);
     }
 }
