+++ CTSTools.WEB/App/Features/Engineering/ComponentID/DecoderManagement/SubClass_Supplier/SubClass_SupplierController.cs
@@ -1,4 +1,5 @@
 ﻿using CTSTools.BLL.Common;
+using CTSTools.BLL.Features.AdvancedSettings.SecurityManagement.Action;
 using CTSTools.BLL.Features.Engineering.ComponentID.DecoderManagement.SubClass_Supplier;
 using CTSTools.WEB.App_Start;
 using DevExtreme.AspNet.Data;
@@ -45,25 +46,35 @@ public class SubClass_SupplierController : ApiController
     [Route("api/SubClass_Supplier/Create")]
     public IHttpActionResult CreateSubClass_Supplier([FromBody] SubClass_SupplierDTO SubClass_SupplierDTO)
     {
-
-        SubClass_SupplierDTO.AddedByID = Auth_Helper.GetLoggedUserOid();
-        var _validationResultDTO = SubClass_Supplier_Service.CreateSubClass_Supplier_Global(SubClass_SupplierDTO);
+        var _validationResultDTO = Auth_Helper.ValidatePermission_Global(nameof(SubClass_Supplier), (int)Action_Enum.Create);
+        if (_validationResultDTO.Result)
+        {
+            SubClass_SupplierDTO.AddedByID = Auth_Helper.GetLoggedUserOid();
+            _validationResultDTO = SubClass_Supplier_Service.CreateSubClass_Supplier_Global(SubClass_SupplierDTO);
+        }
         return Json(_validationResultDTO);
     }
     [HttpPost]
     [Route("api/SubClass_Supplier/Update")]
     public IHttpActionResult UpdateSubClass_Supplier([FromBody] SubClass_SupplierDTO SubClass_SupplierDTO)
     {
-
-        SubClass_SupplierDTO.LastUpdateByID = Auth_Helper.GetLoggedUserOid();
-        var _validationResultDTO = SubClass_Supplier_Service.UpdateSubClass_Supplier_Global(SubClass_SupplierDTO);
+        var _validationResultDTO = Auth_Helper.ValidatePermission_Global(nameof(SubClass_Supplier), (int)Action_Enum.Update);
+        if (_validationResultDTO.Result)
+        {
+            SubClass_SupplierDTO.LastUpdateByID = Auth_Helper.GetLoggedUserOid();
+            _validationResultDTO = SubClass_Supplier_Service.UpdateSubClass_Supplier_Global(SubClass_SupplierDTO);
+        }
         return Json(_validationResultDTO);
     }
     [HttpPost]
     [Route("api/SubClass_Supplier/Delete")]
     public IHttpActionResult DeleteSubClass_Supplier([FromBody] SubClass_SupplierDTO SubClass_SupplierDTO)
     {
-        var _validationResultDTO = SubClass_Supplier_Service.DeleteSubClass_Supplier_Global(SubClass_SupplierDTO);
+        var _validationResultDTO = Auth_Helper.ValidatePermission_Global(nameof(SubClass_Supplier), (int)Action_Enum.Delete);
+        if (_validationResultDTO.Result)
+        {
+            _validationResultDTO = SubClass_Supplier_Service.DeleteSubClass_Supplier_Global(SubClass_SupplierDTO);
+        }
         return Json(_validationResultDTO);
     }
 }
